requirements.txt
django==5.2.14
djangorestframework==3.17.1
